Deprecate the fab module in favour of the util one
